+++ Wabisabi/src/Platform/Opengl/OpenglVertexArray.h
@@ -11,7 +11,7 @@ namespace Wabisabi
 		virtual void Bind() const override;
 		virtual void UnBind() const override;
 		virtual void AddVertexBuffer(std::shared_ptr<VertexBuffer>& buffer) override;
-		virtual void AddVertexBuffer(const void* data,uint32_t size, BufferLayout& layout) override;
+		virtual void AddVertexBuffer(const void* data,size_t size, BufferLayout& layout) override;
 		virtual void SetIndexBuffer(const std::shared_ptr<IndexBuffer>& buffer) override;
 		virtual void SetIndexBuffer(const uint32_t* data, uint32_t count) override;
 		inline virtual const std::shared_ptr<IndexBuffer>& GetIndexBuffer() const override { return m_IndexBuffer; };
